Code cleanup, fix images, javadocs
